(1) Refactor, the logo of the app (2) Refactor, some scale and styling
