+++ User-Frontend/src/components/HomeScreen/HomeScreen.css
@@ -12,6 +12,7 @@
 /* Header Row */
 .header-row {
   display: flex;
+  flex-direction: column;
   justify-content: space-between;
   align-items: center;
   margin-bottom: 1.2rem;
@@ -22,14 +23,20 @@
 
 /* Left: Logo */
 .logo-col {
-  flex: 1;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  padding-bottom: 10px;
 }
 
 .logo {
-  width: 60px;
-  height: 60px;
+  width: 80%;
+  height: 80%;
+  background-color: #050101;
+  padding: 10px;
   border-radius: 10px;
-  object-fit: contain;
+  object-fit: fill;
   box-shadow: 0 2px 6px rgba(0, 0, 0, 0.3);
 }
