+++ User-Frontend/src/components/LoginSignup/LoginSignup.css
@@ -19,13 +19,15 @@ body {
   display: flex;
   justify-content: center;
   align-items: center;
-  margin-bottom: 1rem;
+  margin-bottom: 3.5rem;
 }
 
 .login-logo {
-  width: 80px;
-  height: 80px;
-  object-fit: contain;
+  width: 220px;
+  height: 100px;
+  background-color: #050101;
+  padding: 10px;
+  object-fit: fill;
   border-radius: 12px;
   box-shadow: 0 3px 12px rgba(0, 0, 0, 0.1);
 }
