+++ User-Frontend/src/components/Search/SearchBar.css
@@ -3,6 +3,7 @@
   align-items: center;
   justify-content: center;
   width: 100%;
+  height: 50%;
   position: relative;
   border-radius: 14px;
   background-color: #fff;
